Move NoQa into rules
