+++ src/sqlfluff/core/linter/__init__.py
@@ -4,13 +4,10 @@ from sqlfluff.core.linter.common import ParsedString, RenderedFile, RuleTuple
 from sqlfluff.core.linter.linted_file import LintedFile
 from sqlfluff.core.linter.linter import Linter
 from sqlfluff.core.linter.linting_result import LintingResult
-from sqlfluff.core.linter.noqa import IgnoreMask, NoQaDirective
 
 __all__ = (
     "RuleTuple",
     "ParsedString",
-    "NoQaDirective",
-    "IgnoreMask",
     "LintedFile",
     "LintingResult",
     "Linter",
